feat: combine deprec incentives and tip_118
